# Export performance optimization