--- a/pkg/manager/jsonschema/mappings.go
+++ b/pkg/manager/jsonschema/mappings.go
@@ -1,8 +1,8 @@
 package jsonschema
 
 import (
-	"encoding/json"
 	"fmt"
+	"github.com/json-iterator/go"
 	"os"
 )
 
@@ -31,6 +31,7 @@ func LoadMappingsFile(filePath string) (*Mappings, error) {
 	if err != nil {
 		return nil, err
 	}
+	var json = jsoniter.ConfigCompatibleWithStandardLibrary
 	jsonParser := json.NewDecoder(file)
 	err = jsonParser.Decode(&mappings)
 	if err != nil {