--- a/pkg/manager/paths/paths_json.go
+++ b/pkg/manager/paths/paths_json.go
@@ -6,6 +6,8 @@ import (
 )
 
 type jsonPaths struct {
+	Metadata string
+
 	MappingsFile string
 	ScrapedFile  string
 
@@ -18,6 +20,7 @@ type jsonPaths struct {
 
 func newJSONPaths() *jsonPaths {
 	jp := jsonPaths{}
+	jp.Metadata = config.GetMetadataPath()
 	jp.MappingsFile = filepath.Join(config.GetMetadataPath(), "mappings.json")
 	jp.ScrapedFile = filepath.Join(config.GetMetadataPath(), "scraped.json")
 	jp.Performers = filepath.Join(config.GetMetadataPath(), "performers")
@@ -26,6 +29,11 @@ func newJSONPaths() *jsonPaths {
 	jp.Studios = filepath.Join(config.GetMetadataPath(), "studios")
 	jp.Movies = filepath.Join(config.GetMetadataPath(), "movies")
 	return &jp
+}
+
+func GetJSONPaths() *jsonPaths {
+	jp := newJSONPaths()
+	return jp
 }
 
 func (jp *jsonPaths) PerformerJSONPath(checksum string) string {