Improved the export functionality to utilize multiple threads when creating the json files. 
In my case ( nvme ssd for db and metadata , dual 6c/12t processor) the export function seemed to be cpu bottlenecked due to single thread usage.I suspect the base64 conversion for images might have something to do with that.
I added some worker threads  ( by default GOMAXPROCS in number )
The difference in performance is more evident for the scenes part in my case

~11k scenes ( ~700 mb   in metadata / scenes folder)
_Time to export scenes_
Single thread      **3m33s**
**4**   workers       **50s**
**8**   workers        **26s**
**12** workers       **19s**
**24** workers      **15s**

By default workers  is set as the number of CPUs 
You can experiment with by setting the env GOMAXPROCS 
or by running stash like this (linux 4 workers)
`GOMAXPROCS=4 ./stash-linux`