@WithoutPants 2020-04-18
Seems strange that the bottleneck is the CPU.