@bnkai 2020-04-18
I didn't expect that either. I noticed that a single core ( out of 24) was constantly maxed out during export . My db and metadata were located on nvme ( benches 1GB/s +) and i have plenty of ram available also ( 48gb) so the bottleneck was something else.