@WithoutPants 2020-04-20
In the original code, I profiled the export process, with the following graph:

![pprof](https://user-images.githubusercontent.com/53250216/79721481-bad87d00-8325-11ea-831d-5db736f898e8.png)

I did a quick change to add `FindNamesBySceneID` to `PerformerQueryBuilder` which selects the names of the performers only, and called that from `ExportScenes`. On the database I was testing on, the export went from taking 2.17 minutes to 53 seconds. The graph for that run is below:

![pprof_post](https://user-images.githubusercontent.com/53250216/79722580-a72e1600-8327-11ea-9251-b068ba9ff4c1.png)

I'm guessing it's the blobs in the database that are causing the slowdown (yet another mark for not storing them there).

That was with 10 minutes of investigation or so. With enough time we can probably optimise further. I'd prefer finding and fixing the bottlenecks before we optimise by adding threads.

For profiling I added this code:
```
import "runtime/pprof"

f, _ := os.Create("export.prof")
pprof.StartCPUProfile(f)
defer pprof.StopCPUProfile()

// code to be profiled
```

Then used `go tool pprof -svg stash <path to export.prof> > image.svg` to generate the graph image.