@bnkai 2020-04-20
Are you sure you were cpu bottlenecked and not io ?

I tried changing the part you said and got a nice  speed bump but not that big comapred to yours.
From my test 11k files 536 performers ( NVME ssd)

single worker before change 
![profile01](https://user-images.githubusercontent.com/48220860/79752114-6ed50a80-831c-11ea-9f42-e8b7e6461c71.png)
after change 
![profile01-opt](https://user-images.githubusercontent.com/48220860/79752146-7eecea00-831c-11ea-8303-a921098009cc.png)

there is a ~40 sec gain 4.50 - > 4.10 minutes
As you can see no io threads are present in the graph before/after compared to your profiling.

8 workers
![profile08](https://user-images.githubusercontent.com/48220860/79752889-a3959180-831d-11ea-82dd-068fc4eb30c4.png)

8 workers after change
![profile08-opt](https://user-images.githubusercontent.com/48220860/79752901-a8f2dc00-831d-11ea-9c65-983cef6d841a.png)

35s -> 26s from the change 
4.10 minutes -> 26s  1 worker change - > 8 workers change

I'll try some more profiling to see if i can find something else that explains my cpu bottleneck but in my case threading still gives substancial gains. Can you double check your cpu usage while doing the export  to see if any of your cpu cores is maxed out?