@bnkai 2020-04-20
Lol  FindBySceneID in TagQueryBuilder was using an unnecessary join with scenes.
Removed join and got to 
33s for 1 worker
![profile01-opt2](https://user-images.githubusercontent.com/48220860/79760336-99c55b80-8328-11ea-88a2-8070d2a011fc.png)

and
6.5 s for 8 workers
![profile08-opt](https://user-images.githubusercontent.com/48220860/79760628-04769700-8329-11ea-942e-0c3c6632bd20.png)


It seems @InfiniteTF  has the join removal included in #478  also