@bnkai 2020-04-20
Replaced  `encoding/json` with `github.com/json-iterator/go` in jsonschema 
got from 33s to 21s for 1 worker and from 6.5s to 4.8s using 8 workers
1 worker
![profile01-opt3](https://user-images.githubusercontent.com/48220860/79769189-79030300-8334-11ea-8707-8510f1037d02.png)
8 workers
![profile08-opt3](https://user-images.githubusercontent.com/48220860/79769245-86b88880-8334-11ea-8322-62deb386bab1.png)

with the last change we have 
21s 1 worker
11s 2 workers
6,7s 4 workers
4.8s 8 workers

For my case i still have cpu bottleneck ( 1 core maxed out  ) but now its only for 20s