@WithoutPants 2020-04-21
Exporting when the directory structure hasn't been created results in only `mappings.json` and `scraped.json` being created.