@bnkai 2020-04-21
The directory structure is created once during the setup i think. Did you change that in the settings?