@WithoutPants 2020-04-21
I think it was a problem at my end. I think I must've deleted the directory structure before running the `develop` version, but must've deleted the directory structure while the branch version was running, meaning that it didn't recreate the structure. In fact, if I look at the logs after exporting, I can see errors because the directories don't exist. Regardless, this is certainly nothing to do with your changes.

If we are going to include creating the directory structure as part of the export, then can you combine the repeated code (at startup and during export) into a separate function?