@bnkai 2020-04-22
Deleting the structure while stash was running explains your case.
Working on the extra function now.